bugfix: increase counter range
it was currently only able to count to 255 and start over

+++ src/sleep.ino
@@ -126,7 +126,7 @@ void loop()
   counter++;
   shake_blink();
   enterSleep(6);
-  byte cr = counter;
+  int cr = counter;
   while(cr > 0) {
     blink_signal(1 + (cr % 10));
     enterSleep(7);
